requirements.txt
PyQt6>=6.4.0
yt-dlp>=2023.1.6
requests>=2.28.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
pyinstaller